feat(权限): 任务创建/编辑/关闭收归项目管理员，关闭申请审批退场（#180，裁决 10）
领域与契约：
- CanCreateTask／CanEditTaskConfig／TaskEditRule 收窄为仅可编辑项目者
  （项目负责人与管理员）；裁决 D2（#137）四角色口径作废；任务创建邀请
  （AC-03）保留为成员经管理员授权的间接创建路径；
- 配置输入源／参与人／接收方随 CanEditTaskConfig 收归管理员；负责人保留
  上传链路：交付物项（CanManageDeliverables、DeliverableStructureRule）、
  过程文件／外部材料（CanManageTaskFiles）、成果审核人（CanManageReviewers）
  改为「负责人／项目管理员」（新 OwnerOrProjectAdmin）；
- 关闭申请域整体删除：CancelRoute／DecideCancelRequestRule／
  CanAbandonCancelRequest／CancelRequestStateLabel、mywork 关闭申请审批
  与等待条目、blocker 审批超时的 cancel_request 环节（余成果审核、终审
  两环节）、StartResultUpdateRule 与 PendingApprovalOnTask 去布尔参数；
- 新增 CloseTaskRule／CanCloseTask：项目管理员直接关闭——原因必填、即时
  生效、写任务动态（新动态类型 task_closed；cancel_* 三类仅历史回显）；
- 契约删除 cancel-requests 两端点与 CancelRequest* schema、
  Task.cancelRequest、TaskDetail.cancelRequests、报告 pendingApprovals.
  cancelRequests；/cancellation 改 closeTask（CloseTaskRequest）；两端重生成。

存储与演示数据：
- 迁移 00047：DROP TABLE field_change_requests（#172 后仅存关闭申请一类，
  存量随机制退场清除）；ApplyTaskKeyFields 移至 tasks.sql；
  CountPendingApprovalsByKeyResult 只计待终审完成申请；
- seed：关闭申请数据与动态改为 task_closed 关闭留痕。

前端：
- 抽屉删关闭申请卡片／审批与放弃动作／退回弹窗；「关闭任务」弹窗改
  管理员直接关闭文案；任务列表状态备注去关闭申请分支；我的工作徽标表、
  报告待决策段随契约收窄。

文档：主 PRD §0.19 带票号＋§0.1 决策 1/3、§3.4 权限矩阵重写、§4.1、
§5.1／5.2、§9.1、AC-23/57/60/61；我的工作 PRD V1.1.3 修订记录＋职责／
分组／判定图／通知表，MW-06 作废；CONTEXT.md「关闭申请（已退役）」等词条。

验证：server go build/vet/test 全绿（迁移至 47，集成测试改写为管理员
直接关闭口径）；web build＋tsc＋vitest 24/24；Playwright e2e 74/75＋
复跑失败文件 3/3（原失败为登录填充偶发超时）；8080/5173 无残留进程。

+++ server/internal/api/handler_okr_edit.go
@@ -190,7 +190,7 @@ func (s *Server) UpdateKeyResult(w http.ResponseWriter, r *http.Request, project
 		writeInternalError(w, r, err)
 		return
 	}
-	pending := int32(0)
+	pending := int64(0)
 	if handover {
 		// 未决审批单跟着 KR 负责人走：审批人本就是「所属 KR 负责人」这一职责，
 		// 不转交就得让离任者继续处理，转交与否由发起人在确认框里定（AC-61）。
